Replace jQuery-style getters/setters with actual getters/setters...
Having one method which either gets or sets depending on what arguments
it was passed is intuitive enough for straightforward manual uses (i.e.
when you're directly writing person.name("Bob") or person.age()), but is
actually fairly semantically weak.  Javascript methods can always be
called with arbitrary arguments, and for a nondestructive method to
suddenly become destructive because it unexpectedly got passed an
argument is dangerous.  This is particularly relevant when attributes
are being accessed from generated code, such as compiled Handlebars
templates.

The more fundamental reason this is a problem is that different
libraries use different conventions for how to use Javascript's flexible
argument passing.  jQuery-style accessors use the presence or absence of
an argument as a simple form of method name overloading: i.e. to have
completely different behaviours with the same name, depending on whether
arguments were passed.

Handlebars, however, assumes the functions it calls in view objects are
helpers with optional arguments, so it always passes in the rendering
context just in case the helper wants it.

The "optional arguments" approach makes fewer assumptions than the
"method name overloading", since methods which don't need the optional
arguments simply ignore them, and those which might need them can be
defensive and check whether they're undefined.

Javascript getters and setters play better with generated code, and IMHO
the setters are clearer to read than jQuery-style anyway (because
assignment looks like assignment).  They're supported in all major
browsers except IE.